implemented 2 React Routes, homepage and app
