+++ client/pages/appPage.jsx
@@ -1,0 +1,9 @@
+import React, { useEffect, useState } from 'react';
+
+export default function appPage() {
+  return (
+    <div>
+      <h1>appPage</h1>
+    </div>
+  )
+}
