+++ package.json
@@ -31,6 +31,7 @@
     "react": "^16.14.0",
     "react-devtools": "^4.10.1",
     "react-dom": "^16.14.0",
+    "react-hooks": "^1.0.1",
     "react-router-dom": "^5.2.0",
     "request": "^2.79.0"
   },
@@ -48,6 +49,7 @@
     "html-webpack-plugin": "^5.3.1",
     "node-sass": "^5.0.0",
     "nodemon": "^2.0.7",
+    "react-router": "^5.2.0",
     "sass-loader": "^11.0.1",
     "style-loader": "^2.0.0",
     "webpack": "^5.27.2",
